Moved dummy sound interface
