+++ src/main/java/org/runejs/client/sound/Class8_Sub1_Sub1.java
@@ -1,6 +1,5 @@
 package org.runejs.client.sound;
 
-import org.runejs.client.Interface2;
 import org.runejs.client.util.Signlink;
 
 import java.awt.*;
